+++ CHANGELOG.md
@@ -13,6 +13,9 @@ New descriptors:
 * ElementCenter
 * ElementSize
 
+New value objects:
+* Size
+
 
 ## 0.3: Relative Positioning